Enemy can move towards player from a certain distance and can return to it original position

+++ Assets/Scripts/Interface/IDamageable.cs
@@ -2,7 +2,7 @@ using UnityEngine;
 
 public interface IDamageable
 {
-    public int health { get; }
+    public float health { get; }
 
-    public void TakeDamage(int damage) { }
+    public void TakeDamage(float damage) { }
 }
